movie_app_backend/requirements.txt
Django
numpy
pandas
requests
scipy
sklearn
wheel
gunicorn
psycopg2
django-test-without-migrations

# test for authentification from angular
djangorestframework
djangorestframework-jwt
djangorestframework_simplejwt